# public/openapi.yaml
openapi: 3.0.3
info:
  title: E-Puskesmas EIS API
  version: 1.0.0
  description: |
    OpenAPI untuk E-Puskesmas EIS. Disusun mengikuti struktur menu frontend:
    Dashboard, Cari, Klaster 1-4, Lintas Klaster, Monitoring, dan Akun.

    **Integrasi pihak ketiga (e-Puskesmas / webhook)**
    - Semua pengiriman data dari sistem eksternal dilakukan via endpoint `POST /api/eis/klaster1`, `POST /api/eis/klaster2`, `POST /api/eis/klaster3`, `POST /api/eis/klaster4`, dan `POST /api/eis/lintas-klaster`.
    - Endpoint dashboard/analitik tetap `GET` (read-only). Data yang tampil berasal dari hasil POST (webhook push).
    - Semua endpoint POST di atas wajib menyertakan header `X-Epus-Signature` untuk HMAC:
      signature = HMAC_SHA256(EPUS_WEBHOOK_SECRET, `${t}.${rawJsonUtf8}`)
      Header format: `t=<unix_ts>,v1=<hex_sha256>`
    - Idempotency: gunakan `batchId` unik agar retry aman (tidak dobel).
    - Skala besar: payload boleh `Content-Encoding: gzip` (signature dihitung dari JSON *setelah* di-decompress).
servers:
  - url: http://localhost:3000
    description: Local development
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    cookieAuth:
      type: apiKey
      in: cookie
      name: auth_token
  parameters:
    XEpusSignature:
      name: X-Epus-Signature
      in: header
      required: true
      description: "Signature HMAC. Format: `t=<unix_ts>,v1=<hex_sha256>`"
      schema:
        type: string
        example: t=1700000000,v1=0123abcd...
    ContentEncoding:
      name: Content-Encoding
      in: header
      required: false
      description: "Opsional untuk payload besar. Gunakan `gzip` bila body di-compress."
      schema:
        type: string
        enum: [gzip]
...
    PuskesmasId:
      name: puskesmasId
      in: query
      required: false
      schema:
        type: string
      description: "ID puskesmas, gunakan `all` untuk semua"
    Bulan:
      name: bulan
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 12
      description: Bulan (1-12)
    Tahun:
      name: tahun
      in: query
      required: false
      schema:
        type: integer
      description: Tahun (YYYY)
...
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        details:
          type: object
          additionalProperties: true
...
    KlasterGenericResponse:
      type: object
      properties:
        success: { type: boolean }
        data:
          type: object
          additionalProperties: true
...
paths:
...
  /api/eis/klaster2:
    get:
      tags: [Klaster 2]
      summary: Data Klaster 2
      parameters:
        - $ref: '#/components/parameters/PuskesmasId'
        - $ref: '#/components/parameters/Bulan'
        - $ref: '#/components/parameters/Tahun'
      responses:
        '200':
          description: Data klaster 2
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KlasterGenericResponse'
    post:
      tags: [Klaster 2]
      summary: Webhook push data Klaster 2 (ANC, Imunisasi)
      description: |
        Dipakai oleh sistem eksternal (e-Puskesmas) untuk mengirim **data agregat bulanan** Klaster 2.

        Minimal kirim salah satu: `antenatalCare` atau `imunisasi`.
        Server akan **upsert**:
        - ANC: `puskesmas + bulan + tahun`
        - Imunisasi: `puskesmas + jenisImunisasi + bulan + tahun`
      parameters:
        - $ref: '#/components/parameters/XEpusSignature'
        - $ref: '#/components/parameters/ContentEncoding'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                batchId: { type: string }
                puskesmasKode: { type: string }
                puskesmasId: { type: string }
                bulan: { type: integer, minimum: 1, maximum: 12 }
                tahun: { type: integer }
                antenatalCare:
                  type: object
                  required: [k1, k4, target]
                  properties:
                    k1: { type: integer }
                    k4: { type: integer }
                    target: { type: integer }
                imunisasi:
                  type: array
                  items:
                    type: object
                    required: [jenisImunisasi, kategori, sasaran, capaian]
                    properties:
                      jenisImunisasi: { type: string }
                      kategori: { type: string }
                      sasaran: { type: integer }
                      capaian: { type: integer }
              oneOf:
                - required: [batchId, puskesmasKode, bulan, tahun]
                - required: [batchId, puskesmasId, bulan, tahun]
            examples:
              full_anc_imunisasi:
                summary: Kirim ANC + Imunisasi (lengkap)
                value:
                  batchId: EPUS-BUMIAYU-2026-02-K2-0001
                  puskesmasKode: "3329XXXX"
                  bulan: 2
                  tahun: 2026
                  antenatalCare:
                    k1: 120
                    k4: 95
                    target: 140
                  imunisasi:
                    - jenisImunisasi: HB-0
                      kategori: bayi
                      sasaran: 110
                      capaian: 98
                    - jenisImunisasi: BCG
                      kategori: bayi
                      sasaran: 110
                      capaian: 102
                    - jenisImunisasi: MR
                      kategori: baduta
                      sasaran: 105
                      capaian: 90
              only_anc:
                summary: Kirim ANC saja
                value:
                  batchId: EPUS-BUMIAYU-2026-02-K2-ANC-0001
                  puskesmasKode: "3329XXXX"
                  bulan: 2
                  tahun: 2026
                  antenatalCare:
                    k1: 120
                    k4: 95
                    target: 140
              only_imunisasi:
                summary: Kirim Imunisasi saja
                value:
                  batchId: EPUS-BUMIAYU-2026-02-K2-IMUN-0001
                  puskesmasKode: "3329XXXX"
                  bulan: 2
                  tahun: 2026
                  imunisasi:
                    - jenisImunisasi: HB-0
                      kategori: bayi
                      sasaran: 110
                      capaian: 98
      responses:
        '200':
          description: Data diterima dan diproses
          content:
            application/json:
              schema: { type: object, additionalProperties: true }
        '400':
          description: Validasi gagal
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '401':
          description: Signature invalid
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '500':
          description: Kesalahan server
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
...